LunchVote/requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
black==23.7.0
click==8.1.6
Django==4.2.3
django-admin==2.0.2
django-apscheduler==0.6.2
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
flake8==6.0.0
isoweek==1.3.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2023.3
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tomli==2.0.1
typing_extensions==4.7.1
tzlocal==5.0.1
xlwt==1.3.0
